requirements.txt
Flask
psycopg2-binary
bcrypt